tcp: enable CC algorithms to use ECT(1)
When ECN is successfully negociated for a TCP flow, it defaults to
always use ECT(0) in the IP header.

This patch enables congestion control algorithms to control whether
ECT(0) or ECT(1) should be used on a per-segment basis.

- A new CA module flag (TCP_CONG_WANTS_ECT_1) defines the behavior
  expected by the CA when not-yet initialized for the connection.
  As such, it implicitely assumes that the CA also has the
  TCP_CONG_NEEDS_ECN set. tcp_ca_wants_ect_1() is a new helper
  function to check this flag.
- TCP_ECN_ECT_1 is a new member for (struct tcp_sock*)->ecn_flags,
  and signals whether the connection (currently) wants to use ECT(1)
  or not.

Applying the mark itself on the ToS byte is done in a new helper
function (__INET_ECN_XMIT()) which checks its second argument to
apply ECT(0) or ECT(1). INET_ECN_XMIT() default to the current
behavior, i.e.,using ECT(0)

Signed-off-by: Olivier Tilmans <olivier.tilmans@nokia-bell-labs.com>