net: TCP-Prague tracks alpha in segments
We do not have AccECN byte counters, hence had to rely
on estimating the average packet size to update alpha in
ECN marked bytes. We could have gone for tp->mss_cache as
well.

Tracking only the number of segments is at worst a safe
guess of the congestion level (i.e., if we once had a
sub-mss segment). It is thus a good simplification.