net: safer GSO sizing for TCP Prague
If we output part of our cwnd at a higher rate, decrease
accordingly the size of the GSO burst to ensure we do
not cause a transient queue at the bottleneck.

In the future, max_gso_burst could/should be tracked by
a slower EWMA that tries to gradually rise it unless it
negatively affects alpha.